Implement account registration #14
- Introduced initial SQL helper integration
- Added Envelope, IsAlphaNumeric, Without string extensions
- Added GetDefault type extension
- Rename Client>NetClient Server>NetServer to reduce conflict issues and match NetController
- NetController now properly supplies packet ID to NetEventArgs
- Added serializing null values in Packet. Only applicable for strings at this time.
- Fix Packet.Serialize now writes default values. Caused partial packets to throw when read. This intentionally allows packets to be incomplete; any nulls will attempt to serialize as a default value.
- Added null handling for Packet.WriteString
- Updated PacketManager to create packet IDs as a ushort.
- Introduced Security utils for salts and hashes

+++ Library/Networking/PacketManager.cs
@@ -73,7 +73,7 @@ namespace Swordfish.Library.Networking
                 {
                     if (typeof(ISerializedPacket).IsAssignableFrom(type))
                     {
-                        int id = packetAttribute.PacketID ?? type.FullName.ToSeed();
+                        ushort id = (ushort)(packetAttribute.PacketID ?? type.FullName.ToSeed());
                         PacketDefinition definition = new PacketDefinition {
                             ID = id,
                             Type = type,
